+++ Sources/Sentry/SentrySpan.m
@@ -19,6 +19,11 @@
 #import "SentryTraceHeader.h"
 #import "SentryTracer.h"
 
+#if SENTRY_HAS_UIKIT
+#    import <SentryFramesTracker.h>
+#    import <SentryScreenFrames.h>
+#endif // SENTRY_HAS_UIKIT
+
 NS_ASSUME_NONNULL_BEGIN
 
 @interface
@@ -30,9 +35,18 @@ SentrySpan ()
     NSMutableDictionary<NSString *, id> *_tags;
     NSObject *_stateLock;
     BOOL _isFinished;
+#if SENTRY_HAS_UIKIT
+    NSUInteger initTotalFrames;
+    NSUInteger initSlowFrames;
+    NSUInteger initFrozenFrames;
+    SentryFramesTracker *_framesTracker;
+#endif // SENTRY_HAS_UIKIT
 }
 
 - (instancetype)initWithContext:(SentrySpanContext *)context
+#if SENTRY_HAS_UIKIT
+                  framesTracker:(nullable SentryFramesTracker *)framesTracker;
+#endif // SENTRY_HAS_UIKIT
 {
     if (self = [super init]) {
         self.startTimestamp = [SentryDependencyContainer.sharedInstance.dateProvider date];
@@ -43,6 +57,17 @@ SentrySpan ()
 
         if ([NSThread isMainThread]) {
             _data[SPAN_DATA_THREAD_NAME] = @"main";
+
+#if SENTRY_HAS_UIKIT
+            // Only track frames if running on main thread.
+            _framesTracker = framesTracker;
+            if (_framesTracker.isRunning) {
+                SentryScreenFrames *currentFrames = _framesTracker.currentFrames;
+                initTotalFrames = currentFrames.total;
+                initSlowFrames = currentFrames.slow;
+                initFrozenFrames = currentFrames.frozen;
+            }
+#endif // SENTRY_HAS_UIKIT
         } else {
             _data[SPAN_DATA_THREAD_NAME] = [SentryDependencyContainer.sharedInstance.threadInspector
                 getThreadName:currentThread];
@@ -64,9 +89,17 @@ SentrySpan ()
     return self;
 }
 
-- (instancetype)initWithTracer:(SentryTracer *)tracer context:(SentrySpanContext *)context
+- (instancetype)initWithTracer:(SentryTracer *)tracer
+                       context:(SentrySpanContext *)context
+#if SENTRY_HAS_UIKIT
+                 framesTracker:(nullable SentryFramesTracker *)framesTracker
+{
+    if (self = [self initWithContext:context framesTracker:framesTracker]) {
+#else
 {
     if (self = [self initWithContext:context]) {
+#endif // SENTRY_HAS_UIKIT
+
         _tracer = tracer;
     }
     return self;
@@ -171,6 +204,27 @@ SentrySpan ()
         SENTRY_LOG_DEBUG(@"Setting span timestamp: %@ at system time %llu", self.timestamp,
             (unsigned long long)SentryDependencyContainer.sharedInstance.dateProvider.systemTime);
     }
+
+#if SENTRY_HAS_UIKIT
+    if (_framesTracker.isRunning) {
+
+        SentryScreenFrames *currentFrames = _framesTracker.currentFrames;
+        NSInteger totalFrames = currentFrames.total - initTotalFrames;
+        NSInteger slowFrames = currentFrames.slow - initSlowFrames;
+        NSInteger frozenFrames = currentFrames.frozen - initFrozenFrames;
+
+        if (sentryShouldAddSlowFrozenFramesData(totalFrames, slowFrames, frozenFrames)) {
+            [self setDataValue:@(totalFrames) forKey:@"frames.total"];
+            [self setDataValue:@(slowFrames) forKey:@"frames.slow"];
+            [self setDataValue:@(frozenFrames) forKey:@"frames.frozen"];
+
+            SENTRY_LOG_DEBUG(@"Frames for span \"%@\" Total:%ld Slow:%ld Frozen:%ld",
+                self.operation, (long)totalFrames, (long)slowFrames, (long)frozenFrames);
+        }
+    }
+
+#endif // SENTRY_HAS_UIKIT
+
     if (self.tracer == nil) {
         SENTRY_LOG_DEBUG(
             @"No tracer associated with span with id %@", self.spanId.sentrySpanIdString);