+++ Sources/Sentry/include/SentrySpan.h
@@ -6,6 +6,10 @@ NS_ASSUME_NONNULL_BEGIN
 @class SentryTracer, SentryId, SentrySpanId, SentryFrame, SentrySpanContext;
 @protocol SentrySerializable;
 
+#if SENTRY_HAS_UIKIT
+@class SentryFramesTracker;
+#endif // SENTRY_HAS_UIKIT
+
 @interface SentrySpan : NSObject <SentrySpan, SentrySerializable>
 SENTRY_NO_INIT
 
@@ -85,13 +89,22 @@ SENTRY_NO_INIT
  * @param transaction The @c SentryTracer managing the transaction this span is associated with.
  * @param context This span context information.
  */
-- (instancetype)initWithTracer:(SentryTracer *)transaction context:(SentrySpanContext *)context;
+- (instancetype)initWithTracer:(SentryTracer *)transaction
+                       context:(SentrySpanContext *)context
+#if SENTRY_HAS_UIKIT
+                 framesTracker:(nullable SentryFramesTracker *)framesTracker;
+#endif // SENTRY_HAS_UIKIT
+;
 
 /**
  * Init a @c SentrySpan with given context.
  * @param context This span context information.
  */
-- (instancetype)initWithContext:(SentrySpanContext *)context;
+- (instancetype)initWithContext:(SentrySpanContext *)context
+#if SENTRY_HAS_UIKIT
+                  framesTracker:(nullable SentryFramesTracker *)framesTracker;
+#endif // SENTRY_HAS_UIKIT
+;
 
 - (void)setExtraValue:(nullable id)value forKey:(NSString *)key DEPRECATED_ATTRIBUTE;
 @end